\name{cfGAL}
\alias{cfGAL}
\title{Characteristic function of GAL distribution}
\usage{
cfGAL(u, param, type = "mu", log = FALSE)
}
\arguments{
  \item{u}{Transform variate}

  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Characteristic function value at point u for given
parameter vector
}
\description{
Characteristic function of GAL distribution
}
\author{
Francois Pelletier
}

