\name{cgfEsscherGAL}
\alias{cgfEsscherGAL}
\title{Cumulant generating function of the
Esscher transform with parameter 1 of GAL distribution}
\usage{
cgfEsscherGAL(u, param, eval.time = 1, type = "mu", log = FALSE)
}
\arguments{
  \item{u}{Transform variate}

  \item{param}{Parameter vector}

  \item{eval.time}{Time of the process}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Cumulant generating function value at point u for given
parameter vector
}
\description{
Cumulant generating function of the Esscher transform with
parameter 1 of GAL distribution
}
\author{
Francois Pelletier
}

