\name{cgfGAL}
\alias{cgfGAL}
\title{Cumulant generating function of GAL distribution}
\usage{
cgfGAL(u, param, type = "mu", log = FALSE)
}
\arguments{
  \item{u}{Transform variate}

  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Cumulant generating function value at point u for given
parameter vector
}
\description{
Cumulant generating function of GAL distribution
}
\author{
Francois Pelletier
}

