\name{changetypeGAL}
\alias{changetypeGAL}
\title{Change type of parametrization of GAL distribution}
\usage{
changetypeGAL(param, type = "mu", target = "kappa", log = FALSE)
}
\arguments{
  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{target}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
The converted parameter vector
}
\description{
Change type of parametrization of GAL distribution
}
\author{
Francois Pelletier
}

