\name{dchangetypeGAL}
\alias{dchangetypeGAL}
\title{Derivative vector of the type change function}
\usage{
dchangetypeGAL(param, type = "mu", target = "kappa", log = FALSE)
}
\arguments{
  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{target}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
The derivative matrix of the type change function
}
\description{
Derivative vector of the type change function
}
\author{
Francois Pelletier
}

