\name{dGAL}
\alias{dGAL}
\title{Density function for the GAL distribution}
\usage{
dGAL(x, param, type = "mu", log = FALSE)
}
\arguments{
  \item{x}{vector of quantiles}

  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
density at quantile x
}
\description{
Density function for the GAL distribution
}
\author{
Francois Pelletier
}

