\name{diffcgfEsscherGAL}
\alias{diffcgfEsscherGAL}
\title{Differenciation of the cumulant generating fonction of the
Esscher transform with parameter 1 of the GAL distribution}
\usage{
diffcgfEsscherGAL(u, order, param, eval.time = 1, type = "mu",
  log = FALSE)
}
\arguments{
  \item{u}{Transform variate point of evaluation}

  \item{order}{Order of differenciation}

  \item{param}{Parameters of the GAL distirbution}

  \item{eval.time}{Time of the process}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
The value of the derivative at the transform variate point
of evaluation
}
\description{
Differenciation of the cumulant generating fonction of the
Esscher transform with parameter 1 of the GAL distribution
}
\author{
Francois Pelletier
}

