\name{dsaddleapproxGAL}
\alias{dsaddleapproxGAL}
\title{Saddlepoint approximation of the density function of the
GAL distribution}
\usage{
dsaddleapproxGAL(x, param, eval.time = 1, type = "mu", log = FALSE)
}
\arguments{
  \item{x}{vector of quantiles}

  \item{param}{Parameter vector}

  \item{eval.time}{Time of the process}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Saddlepoint approximation of the density function
}
\description{
Saddlepoint approximation of the density function of the
GAL distribution
}

