\name{kurtosisGAL}
\alias{kurtosisGAL}
\title{(Adjusted) kurtosis of the GAL distribution}
\usage{
kurtosisGAL(param, type = "mu", log = FALSE, adjust = TRUE)
}
\arguments{
  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}

  \item{adjust}{Logical to use the adjusted kurtosis}
}
\value{
A numeric value of the kurtosis
}
\description{
(Adjusted) kurtosis of the GAL distribution
}
\author{
Francois Pelletier
}

