\name{mean.variance.GMM.gradient}
\alias{mean.variance.GMM.gradient}
\title{Gradient of the GMM moment conditions on mean and variance}
\usage{
\method{mean}{variance.GMM.gradient}(param, Data, type = "mu")
}
\arguments{
  \item{param}{GAL parameters}
}
\value{
gradient matrix
}
\description{
Derivative of the moment conditions according to the vector
of parameters
}
\author{
François Pelletier
}

