\name{mgfGAL}
\alias{mgfGAL}
\title{Moment generating function of GAL distribution}
\usage{
mgfGAL(u, param, type = "mu", log = FALSE)
}
\arguments{
  \item{u}{Transform variate}

  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Moment generating function value at point u for given
parameter vector
}
\description{
Moment generating function of GAL distribution
}
\author{
Francois Pelletier
}

