\name{pGAL}
\alias{pGAL}
\title{Distribution function for the GAL distribution}
\usage{
pGAL(x, param, type = "mu", log = FALSE)
}
\arguments{
  \item{x}{vector of quantiles}

  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
distribution at quantile x
}
\description{
Distribution function for the GAL distribution
}
\author{
Francois Pelletier
}

