\name{pnormapproxEsscherLM}
\alias{pnormapproxEsscherLM}
\title{Normal approximation of the distribution function of the
Esscher transform of a Laplace Motion}
\usage{
pnormapproxEsscherLM(x, param, hEsscher = 0, eval.time = 1, type = "mu",
  log = FALSE, start = 0)
}
\arguments{
  \item{x}{vector of quantiles}

  \item{param}{Parameter vector}

  \item{hEsscher}{Esscher transform parameter}

  \item{eval.time}{Time of the process}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}

  \item{start}{Starting value of the process}
}
\value{
Normal distribution function approximation
}
\description{
Normal approximation of the distribution function of the
Esscher transform of a Laplace Motion
}
\author{
Francois Pelletier
}

