\name{psaddleapproxEsscherGAL}
\alias{psaddleapproxEsscherGAL}
\title{Saddlepoint approximation of the distribution function of the Esscher
transform of the GAL distribution}
\usage{
psaddleapproxEsscherGAL(x, param, eval.time = 1, type = "mu", log = FALSE)
}
\arguments{
  \item{x}{vector of quantiles}

  \item{param}{Parameter vector}

  \item{eval.time}{Time of the process}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Saddlepoint approximation of the distribution function
}
\description{
Saddlepoint approximation of the distribution function of
the Esscher transform of the GAL distribution
}
\author{
Francois Pelletier
}

