\name{psaddleapproxGAL}
\alias{psaddleapproxGAL}
\title{Saddlepoint approximation of the distribution function of the
GAL distribution}
\usage{
psaddleapproxGAL(x, param, eval.time = 1, type = "mu", log = FALSE)
}
\arguments{
  \item{x}{vector of quantiles}

  \item{param}{Parameter vector}

  \item{eval.time}{Time of the process}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Saddlepoint approximation of the distribution function
}
\description{
Saddlepoint approximation of the distribution function of
the GAL distribution
}

