\name{rGAL}
\alias{rGAL}
\title{Random number generator for the GAL distribution}
\usage{
rGAL(n, param, type = "mu", log = FALSE)
}
\arguments{
  \item{n}{number of observations}

  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
A vector of random numbers
}
\description{
Random number generator for the GAL distribution
}
\author{
Francois Pelletier
}

