\name{riskneutralparGAL}
\alias{riskneutralparGAL}
\title{Risk neutral conversion of parameters of GAL distirbution}
\usage{
riskneutralparGAL(param, riskfree, type = "mu", log = FALSE)
}
\arguments{
  \item{param}{Parameter vector}

  \item{riskfree}{Risk-free rate}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
Risk neutral parameter vector
}
\description{
Risk neutral conversion of parameters of GAL distirbution
}
\author{
Francois Pelletier
}

