\name{skewnessGAL}
\alias{skewnessGAL}
\title{Skewness of the GAL distribution}
\usage{
skewnessGAL(param, type = "mu", log = FALSE)
}
\arguments{
  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
A numeric value of the skewness
}
\description{
Skewness of the GAL distribution
}
\author{
Francois Pelletier
}

