\name{testparGAL}
\alias{testparGAL}
\title{Check for the validity of a parameter vector. Stop at error.}
\usage{
testparGAL(param, type = "mu", log = FALSE)
}
\arguments{
  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
logical
}
\description{
Check for the validity of a parameter vector. Stop at
error.
}
\author{
Francois Pelletier
}

