\name{cmGAL}
\alias{cmGAL}
\title{Centered moments of the GAL distribution}
\usage{
cmGAL(order, param, type = "mu", log = FALSE)
}
\arguments{
  \item{order}{Order of raw moment}

  \item{param}{Parameter vector}

  \item{type}{Choose between "mu" or "kappa"
  parametrization}

  \item{log}{Logical for log-parameters}
}
\value{
A numeric value of the centered moment
}
\description{
Centered moments of the GAL distribution
}
\author{
Francois Pelletier
}

